Update README.md
The description of what the program does is above. Now to access it. 
STEP 1: You will need to press the green button labeled "code (it's just above this readme). Tap it, and make sure the HTTPS part is highlighted (if not then press HTTPS).  Under that, you will see a link, copy it. 

STEP 2: Now head to vscode and made a new window (so we can access this project). Now that we are in vs code click clone git repository, on the top there should've been a box where you can enter the HTTP link (so just copy it into the box and press enter ).  Now go through your and see where you want to open the folder (try desktop or documents depending on what you prefer). Now select the folder where you want to open it then click select repository location once you find the right folder (if not you can make a new one before you clone the repository (before step 2). 

STEP 3: Make sure you click trust the author as you open it. Now that it's open create a new terminal (on the top left opposite to the control center, there should say 'terminal' click it and make a new terminal. this is on Mac).  After click 'run' right next to the terminal and press 'start debugging'. You know there were no errors when an open-source program pops up like google chrome or safari. 

STEP 4: Open up postman (if you haven't already downloaded so). There should be a 'My workspace' tab (it has a big front and it's on the right side. Above  'My workspace' there should be a '+' button, press it. Now there should be a 'GET' button and next to that enter a request URL where you will copy this into  'https://localhost:7082/new/ball' (that is the URL to access my program). Is it pasted in there? Great, now press send. 

STEP 5: There should be a body just below that with text. Read it, Now add another '/' after the URL link with your input to that question. If your having trouble look at this as an example 'https://localhost:7082/new/ball/will i get my dream house soon'